requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-admin-thumbnails==0.2.9
django-cors-headers==4.4.0
django-unfold==0.46.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
validate-email==1.3
